Fix ISIS-1715 menuOrder not honored
